+++ class_definitions.lua
