--- preInit/init.lua
@@ -1,1 +1,0 @@
-require("preInit.classDefinitions")
