+++ ui/exit_screen/styles.lua
@@ -1,0 +1,4 @@
+---@class ExitScreenStyles
+local M = {}
+
+return M
