+++ ui/init.lua
@@ -3,12 +3,9 @@ local awful = require("awful")
 
 local volume_popup = require("ui.volume_widget_popup")
 
-load_all("ui", {
-  "bar",
-  "utils",
-  "ExitScreen",
-  "notificationCenter",
-})
+require(... .. ".bar")
+require(... .. ".ExitScreen")
+require(... .. ".notificationCenter")
 
 awful.screen.connect_for_each_screen(function(s)
   volume_popup(s)
