--- ui/utils/init.lua
@@ -1,3 +1,0 @@
-load_all("ui.utils", {
-    "ClickableContainer"
-})
