+++ widgets/init.lua
@@ -1,6 +1,7 @@
 --- @class Widgets
 --- @field text fun(args:{}): TextWidget
 return {
-    text = require(... .. '.text'),
-    button = require(... .. ".button")
+  text = require(... .. ".text"),
+  button = require(... .. ".button"),
+  overflow = require(... .. ".overflow"),
 }
